remove unnecessary view source_w/h i32 casts

+++ gm8emulator/src/game/draw.rs
@@ -127,7 +127,7 @@ impl Game {
 
                     let x = inst.x.get().round().to_i32();
                     let y = inst.y.get().round().to_i32();
-                    if view.follow_hborder < (view.source_w / 2) as i32 {
+                    if view.follow_hborder < view.source_w / 2 {
                         let border_left = x - view.follow_hborder;
                         let border_right = x + view.follow_hborder;
                         if border_left < view.source_x {
@@ -136,20 +136,20 @@ impl Game {
                             } else {
                                 view.source_x -= (view.source_x - border_left).min(view.follow_hspeed);
                             }
-                        } else if border_right > (view.source_x + view.source_w as i32) {
+                        } else if border_right > view.source_x + view.source_w {
                             if view.follow_hspeed < 0 {
-                                view.source_x = border_right - view.source_w as i32;
+                                view.source_x = border_right - view.source_w;
                             } else {
                                 view.source_x +=
-                                    (border_right - (view.source_x + view.source_w as i32)).min(view.follow_hspeed);
+                                    (border_right - (view.source_x + view.source_w)).min(view.follow_hspeed);
                             }
                         }
                     } else {
-                        view.source_x = x - (view.source_w / 2) as i32;
+                        view.source_x = x - view.source_w / 2;
                     }
-                    view.source_x = view.source_x.max(0).min(self.room.width - view.source_w as i32);
+                    view.source_x = view.source_x.max(0).min(self.room.width - view.source_w);
 
-                    if view.follow_vborder < (view.source_h / 2) as i32 {
+                    if view.follow_vborder < view.source_h / 2 {
                         let border_top = y - view.follow_vborder;
                         let border_bottom = y + view.follow_vborder;
                         if border_top < view.source_y {
@@ -158,18 +158,18 @@ impl Game {
                             } else {
                                 view.source_y -= (view.source_y - border_top).min(view.follow_vspeed);
                             }
-                        } else if border_bottom > (view.source_y + view.source_h as i32) {
+                        } else if border_bottom > view.source_y + view.source_h {
                             if view.follow_vspeed < 0 {
-                                view.source_y = border_bottom - view.source_h as i32;
+                                view.source_y = border_bottom - view.source_h;
                             } else {
                                 view.source_y +=
-                                    (border_bottom - (view.source_y + view.source_h as i32)).min(view.follow_vspeed);
+                                    (border_bottom - (view.source_y + view.source_h)).min(view.follow_vspeed);
                             }
                         }
                     } else {
-                        view.source_y = y - (view.source_h / 2) as i32;
+                        view.source_y = y - view.source_h / 2;
                     }
-                    view.source_y = view.source_y.max(0).min(self.room.height - view.source_h as i32);
+                    view.source_y = view.source_y.max(0).min(self.room.height - view.source_h);
                 }
             }
         }
